ZO: Part of refactor implementation
