+++ libs/zzz/svgicons/src/icons/Elements/ElectricIcon.tsx
@@ -1,47 +1,46 @@
-import { SvgIcon } from '@mui/material'
+import { createSvgIcon } from '@mui/material'
 
-export const ElectricIcon = () => (
-  <SvgIcon>
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24px"
-      height="24px"
-      viewBox="0 0 24 24"
-      version="1.1"
-    >
-      <defs>
-        <linearGradient
-          id="zzz_electric_icon"
-          gradientUnits="userSpaceOnUse"
-          x1="12.046219"
-          y1="334.813324"
-          x2="12.046219"
-          y2="349.957123"
-          gradientTransform="matrix(1.572331,0,0,1.572331,-7.035129,-526.43756)"
-        >
-          <stop
-            offset="0"
-            style={{ stopColor: 'rgb(0%,45.882353%,100%)', stopOpacity: 1 }}
-          />
-          <stop
-            offset="1"
-            style={{
-              stopColor: 'rgb(23.921569%,92.54902%,100%)',
-              stopOpacity: 1,
-            }}
-          />
-        </linearGradient>
-      </defs>
-      <g id="surface1">
-        <path
-          style={{
-            stroke: 'none',
-            fillRule: 'nonzero',
-            fill: 'url(#zzz_electric_icon)',
-          }}
-          d="M 16.703125 1.0625 L 7.796875 2.121094 L 5.121094 11.027344 L 8.769531 10.9375 L 5.8125 22.75 L 17.351562 8.074219 L 12.644531 8.445312 Z M 2.722656 9.828125 L 0 13.105469 L 1.152344 17.488281 L 2.261719 16.519531 L 3.875 20.765625 L 4.890625 13.058594 L 2.90625 14.441406 Z M 18.320312 11.535156 L 13.84375 15.691406 L 15.828125 16.195312 L 11.238281 21.964844 L 21.550781 17.257812 L 18.550781 15.921875 L 23.8125 11.90625 Z M 18.320312 11.535156 "
+export const ElectricIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    version="1.1"
+  >
+    <defs>
+      <linearGradient
+        id="zzz_electric_icon"
+        gradientUnits="userSpaceOnUse"
+        x1="12.046219"
+        y1="334.813324"
+        x2="12.046219"
+        y2="349.957123"
+        gradientTransform="matrix(1.572331,0,0,1.572331,-7.035129,-526.43756)"
+      >
+        <stop
+          offset="0"
+          style={{ stopColor: 'rgb(0%,45.882353%,100%)', stopOpacity: 1 }}
         />
-      </g>
-    </svg>
-  </SvgIcon>
+        <stop
+          offset="1"
+          style={{
+            stopColor: 'rgb(23.921569%,92.54902%,100%)',
+            stopOpacity: 1,
+          }}
+        />
+      </linearGradient>
+    </defs>
+    <g id="surface1">
+      <path
+        style={{
+          stroke: 'none',
+          fillRule: 'nonzero',
+          fill: 'url(#zzz_electric_icon)',
+        }}
+        d="M 16.703125 1.0625 L 7.796875 2.121094 L 5.121094 11.027344 L 8.769531 10.9375 L 5.8125 22.75 L 17.351562 8.074219 L 12.644531 8.445312 Z M 2.722656 9.828125 L 0 13.105469 L 1.152344 17.488281 L 2.261719 16.519531 L 3.875 20.765625 L 4.890625 13.058594 L 2.90625 14.441406 Z M 18.320312 11.535156 L 13.84375 15.691406 L 15.828125 16.195312 L 11.238281 21.964844 L 21.550781 17.257812 L 18.550781 15.921875 L 23.8125 11.90625 Z M 18.320312 11.535156 "
+      />
+    </g>
+  </svg>,
+  'ElectricIcon'
 )
