+++ libs/zzz/svgicons/src/icons/Elements/FrostIcon.tsx
@@ -1,71 +1,70 @@
-import { SvgIcon } from '@mui/material'
+import { createSvgIcon } from '@mui/material'
 
-export const FrostIcon = () => (
-  <SvgIcon>
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24px"
-      height="24px"
-      viewBox="0 0 24 24"
-      version="1.1"
-    >
-      <defs>
-        <linearGradient
-          id="zzz_frost_gradient"
-          gradientUnits="userSpaceOnUse"
-          x1="10.739503"
-          y1="1.241672"
-          x2="3.260498"
-          y2="12.758328"
-          gradientTransform="matrix(1.700787,0,0,1.700787,0,0)"
-        >
-          <stop
-            offset="0"
-            style={{
-              stopColor: 'rgb(15.686275%,98.431373%,100%)',
-              stopOpacity: 1,
-            }}
-          />
-          <stop
-            offset="0.227626"
-            style={{
-              stopColor: 'rgb(12.941176%,97.254902%,97.254902%)',
-              stopOpacity: 1,
-            }}
-          />
-          <stop
-            offset="0.5"
-            style={{
-              stopColor: 'rgb(23.529412%,83.921569%,94.509804%)',
-              stopOpacity: 1,
-            }}
-          />
-          <stop
-            offset="0.791829"
-            style={{
-              stopColor: 'rgb(45.098039%,63.529412%,100%)',
-              stopOpacity: 1,
-            }}
-          />
-          <stop
-            offset="1"
-            style={{
-              stopColor: 'rgb(45.882353%,63.921569%,100%)',
-              stopOpacity: 1,
-            }}
-          />
-        </linearGradient>
-      </defs>
-      <g id="surface1">
-        <path
+export const FrostIcon = createSvgIcon(
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    version="1.1"
+  >
+    <defs>
+      <linearGradient
+        id="zzz_frost_gradient"
+        gradientUnits="userSpaceOnUse"
+        x1="10.739503"
+        y1="1.241672"
+        x2="3.260498"
+        y2="12.758328"
+        gradientTransform="matrix(1.700787,0,0,1.700787,0,0)"
+      >
+        <stop
+          offset="0"
           style={{
-            stroke: 'none',
-            fillRule: 'nonzero',
-            fill: 'url(#zzz_frost_gradient)',
+            stopColor: 'rgb(15.686275%,98.431373%,100%)',
+            stopOpacity: 1,
           }}
-          d="M 11.941406 0 C 11.753906 0 11.09375 2.441406 9.644531 3.785156 C 9.585938 3.839844 9.640625 3.992188 9.640625 4.027344 C 10.570312 4.496094 11.394531 5.152344 11.933594 5.96875 C 12.476562 4.992188 13.25 4.375 14.164062 3.972656 C 14.164062 3.972656 14.226562 3.777344 14.164062 3.722656 C 12.785156 2.519531 12.128906 0 11.941406 0 Z M 0.226562 3.480469 C 0.414062 3.890625 0.746094 4.085938 1.046875 4.410156 C 1.402344 4.796875 1.710938 5.292969 1.957031 5.753906 C 2.609375 6.957031 2.875 8.355469 3.679688 9.476562 C 5.097656 11.453125 8.1875 12.40625 10.457031 11.335938 L 9.742188 10.5 L 8.703125 9.0625 C 9.75 9.488281 10.953125 10.636719 11.894531 10.628906 C 12.972656 10.617188 13.984375 9.472656 15.003906 9.0625 C 14.832031 9.613281 14.414062 10.210938 13.972656 10.589844 C 13.792969 10.742188 13.160156 11.261719 13.160156 11.261719 C 13.160156 11.261719 13.367188 11.605469 14.800781 11.640625 C 16.125 11.71875 17.867188 11.625 18.933594 10.707031 C 20.410156 9.429688 20.765625 7.804688 21.644531 6.167969 C 21.949219 5.605469 22.265625 5.019531 22.691406 4.539062 C 22.984375 4.203125 23.34375 3.957031 23.582031 3.585938 C 23.070312 3.398438 22.355469 3.570312 21.828125 3.703125 C 20.503906 4.03125 19.457031 4.4375 18.3125 5.191406 C 17.996094 5.402344 17.144531 5.976562 17.015625 5.238281 C 16.925781 4.726562 17.734375 4.242188 17.898438 3.789062 C 15.03125 4.078125 12.648438 4.523438 11.898438 7.679688 C 11.527344 5.210938 8.285156 3.675781 5.808594 3.789062 C 6.074219 4.207031 7.269531 4.914062 6.6875 5.511719 C 6.277344 5.933594 5.4375 5.113281 5.085938 4.910156 C 3.585938 4.046875 1.980469 3.488281 0.226562 3.480469 Z M 8.074219 12.074219 C 6.984375 12.089844 5.882812 12.316406 5.085938 12.832031 C 4.511719 13.207031 3.917969 13.875 3.535156 14.433594 C 2.8125 15.492188 2.566406 16.824219 1.957031 17.949219 C 1.6875 18.453125 1.339844 18.976562 0.941406 19.394531 C 0.664062 19.683594 0.371094 19.839844 0.226562 20.222656 C 1.953125 20.222656 3.738281 19.671875 5.1875 18.742188 C 5.5 18.542969 6.351562 17.839844 6.671875 18.269531 C 7.070312 18.804688 6.144531 19.410156 6.015625 19.914062 C 8.363281 19.792969 11.1875 19.042969 11.699219 16.398438 C 12.324219 16.824219 12.4375 17.699219 12.949219 18.257812 C 13.886719 19.28125 15.265625 19.535156 16.554688 19.746094 C 17.007812 19.820312 17.570312 20.023438 18.003906 19.914062 C 17.769531 19.550781 16.742188 18.976562 16.949219 18.472656 C 17.21875 17.816406 17.960938 18.378906 18.3125 18.589844 C 19.957031 19.578125 21.566406 20.125 23.480469 20.324219 C 23.183594 19.511719 22.289062 18.835938 21.851562 18.050781 C 21.128906 16.753906 20.839844 15.210938 19.921875 14.023438 C 18.320312 11.9375 15.492188 12.089844 13.136719 12.335938 L 14.15625 13.308594 L 15.003906 14.746094 L 12.835938 13.480469 L 11.828125 13.101562 L 10.769531 13.582031 L 8.597656 14.746094 L 9.449219 13.507812 L 10.457031 12.367188 C 9.769531 12.171875 8.925781 12.058594 8.074219 12.074219 Z M 11.824219 17.84375 C 11.28125 18.820312 10.507812 19.433594 9.59375 19.839844 C 9.59375 19.839844 9.53125 20.035156 9.59375 20.089844 C 10.972656 21.289062 11.628906 23.8125 11.816406 23.8125 C 12.007812 23.8125 12.667969 21.371094 14.113281 20.027344 C 14.171875 19.972656 14.121094 19.816406 14.121094 19.785156 C 13.1875 19.3125 12.367188 18.660156 11.824219 17.84375 Z M 11.824219 17.84375 "
         />
-      </g>
-    </svg>
-  </SvgIcon>
+        <stop
+          offset="0.227626"
+          style={{
+            stopColor: 'rgb(12.941176%,97.254902%,97.254902%)',
+            stopOpacity: 1,
+          }}
+        />
+        <stop
+          offset="0.5"
+          style={{
+            stopColor: 'rgb(23.529412%,83.921569%,94.509804%)',
+            stopOpacity: 1,
+          }}
+        />
+        <stop
+          offset="0.791829"
+          style={{
+            stopColor: 'rgb(45.098039%,63.529412%,100%)',
+            stopOpacity: 1,
+          }}
+        />
+        <stop
+          offset="1"
+          style={{
+            stopColor: 'rgb(45.882353%,63.921569%,100%)',
+            stopOpacity: 1,
+          }}
+        />
+      </linearGradient>
+    </defs>
+    <g id="surface1">
+      <path
+        style={{
+          stroke: 'none',
+          fillRule: 'nonzero',
+          fill: 'url(#zzz_frost_gradient)',
+        }}
+        d="M 11.941406 0 C 11.753906 0 11.09375 2.441406 9.644531 3.785156 C 9.585938 3.839844 9.640625 3.992188 9.640625 4.027344 C 10.570312 4.496094 11.394531 5.152344 11.933594 5.96875 C 12.476562 4.992188 13.25 4.375 14.164062 3.972656 C 14.164062 3.972656 14.226562 3.777344 14.164062 3.722656 C 12.785156 2.519531 12.128906 0 11.941406 0 Z M 0.226562 3.480469 C 0.414062 3.890625 0.746094 4.085938 1.046875 4.410156 C 1.402344 4.796875 1.710938 5.292969 1.957031 5.753906 C 2.609375 6.957031 2.875 8.355469 3.679688 9.476562 C 5.097656 11.453125 8.1875 12.40625 10.457031 11.335938 L 9.742188 10.5 L 8.703125 9.0625 C 9.75 9.488281 10.953125 10.636719 11.894531 10.628906 C 12.972656 10.617188 13.984375 9.472656 15.003906 9.0625 C 14.832031 9.613281 14.414062 10.210938 13.972656 10.589844 C 13.792969 10.742188 13.160156 11.261719 13.160156 11.261719 C 13.160156 11.261719 13.367188 11.605469 14.800781 11.640625 C 16.125 11.71875 17.867188 11.625 18.933594 10.707031 C 20.410156 9.429688 20.765625 7.804688 21.644531 6.167969 C 21.949219 5.605469 22.265625 5.019531 22.691406 4.539062 C 22.984375 4.203125 23.34375 3.957031 23.582031 3.585938 C 23.070312 3.398438 22.355469 3.570312 21.828125 3.703125 C 20.503906 4.03125 19.457031 4.4375 18.3125 5.191406 C 17.996094 5.402344 17.144531 5.976562 17.015625 5.238281 C 16.925781 4.726562 17.734375 4.242188 17.898438 3.789062 C 15.03125 4.078125 12.648438 4.523438 11.898438 7.679688 C 11.527344 5.210938 8.285156 3.675781 5.808594 3.789062 C 6.074219 4.207031 7.269531 4.914062 6.6875 5.511719 C 6.277344 5.933594 5.4375 5.113281 5.085938 4.910156 C 3.585938 4.046875 1.980469 3.488281 0.226562 3.480469 Z M 8.074219 12.074219 C 6.984375 12.089844 5.882812 12.316406 5.085938 12.832031 C 4.511719 13.207031 3.917969 13.875 3.535156 14.433594 C 2.8125 15.492188 2.566406 16.824219 1.957031 17.949219 C 1.6875 18.453125 1.339844 18.976562 0.941406 19.394531 C 0.664062 19.683594 0.371094 19.839844 0.226562 20.222656 C 1.953125 20.222656 3.738281 19.671875 5.1875 18.742188 C 5.5 18.542969 6.351562 17.839844 6.671875 18.269531 C 7.070312 18.804688 6.144531 19.410156 6.015625 19.914062 C 8.363281 19.792969 11.1875 19.042969 11.699219 16.398438 C 12.324219 16.824219 12.4375 17.699219 12.949219 18.257812 C 13.886719 19.28125 15.265625 19.535156 16.554688 19.746094 C 17.007812 19.820312 17.570312 20.023438 18.003906 19.914062 C 17.769531 19.550781 16.742188 18.976562 16.949219 18.472656 C 17.21875 17.816406 17.960938 18.378906 18.3125 18.589844 C 19.957031 19.578125 21.566406 20.125 23.480469 20.324219 C 23.183594 19.511719 22.289062 18.835938 21.851562 18.050781 C 21.128906 16.753906 20.839844 15.210938 19.921875 14.023438 C 18.320312 11.9375 15.492188 12.089844 13.136719 12.335938 L 14.15625 13.308594 L 15.003906 14.746094 L 12.835938 13.480469 L 11.828125 13.101562 L 10.769531 13.582031 L 8.597656 14.746094 L 9.449219 13.507812 L 10.457031 12.367188 C 9.769531 12.171875 8.925781 12.058594 8.074219 12.074219 Z M 11.824219 17.84375 C 11.28125 18.820312 10.507812 19.433594 9.59375 19.839844 C 9.59375 19.839844 9.53125 20.035156 9.59375 20.089844 C 10.972656 21.289062 11.628906 23.8125 11.816406 23.8125 C 12.007812 23.8125 12.667969 21.371094 14.113281 20.027344 C 14.171875 19.972656 14.121094 19.816406 14.121094 19.785156 C 13.1875 19.3125 12.367188 18.660156 11.824219 17.84375 Z M 11.824219 17.84375 "
+      />
+    </g>
+  </svg>,
+  'FrostIcon'
 )
